+++ tests/v2/test_06_tasks_lifecycle/test_unit_bash_scripts.py
@@ -187,37 +187,23 @@ def test_template_2(
     assert condition_1 or condition_2
 
 
-def test_template_4_missing_manifest(
-    tmp_path, testdata_path, current_py_version
-):
-    path = tmp_path / "unit_templates"
-    venv_path = path / "venv_miss_manifest"
-    install_string_miss = testdata_path.parent / (
-        "v2/fractal_tasks_fail/missing_manifest/dist/"
-        "fractal_tasks_mock-0.0.1-py3-none-any.whl"
-    )
-    package_name = "fractal-tasks-mock"
+def test_template_4_missing_manifest(tmp_path, current_py_version):
+    venv_path = (tmp_path / "venv_miss_manifest").as_posix()
     execute_command_sync(
         command=f"python{current_py_version} -m venv {venv_path}"
     )
-    execute_command_sync(
-        command=(
-            f"{venv_path}/bin/python{current_py_version} "
-            f"-m pip install {install_string_miss}"
-        )
-    )
     replacements = [
-        ("__PACKAGE_ENV_DIR__", venv_path.as_posix()),
-        ("__PACKAGE_NAME__", package_name),
+        ("__PACKAGE_ENV_DIR__", venv_path),
+        ("__PACKAGE_NAME__", "pip"),
     ]
-    script_path = tmp_path / "4_good.sh"
+    script_path = (tmp_path / "4_good.sh").as_posix()
     customize_template(
         template_name="4_pip_show.sh",
         replacements=replacements,
-        script_path=script_path.as_posix(),
+        script_path=script_path,
     )
     with pytest.raises(RuntimeError) as expinfo:
-        execute_command_sync(command=f"bash {script_path.as_posix()}")
+        execute_command_sync(command=f"bash {script_path}")
     assert "ERROR: manifest path not found" in str(expinfo.value)
 
 
@@ -237,22 +223,6 @@ def test_templates_freeze(tmp_path, current_py_version):
             replacements=[
                 ("__PACKAGE_ENV_DIR__", venv_path.as_posix()),
                 ("__PYTHON__", f"python{current_py_version}"),
-            ],
-            script_dir=tmp_path,
-            logger_name=__name__,
-            prefix="prefix",
-        )
-        _customize_and_run_template(
-            template_filename="2_pip_install.sh",
-            replacements=[
-                ("__PACKAGE_ENV_DIR__", venv_path.as_posix()),
-                ("__INSTALL_STRING__", "pip"),
-                ("__FRACTAL_MAX_PIP_VERSION__", "99"),
-                ("__PINNED_PACKAGE_LIST__", ""),
-                (
-                    "__FRACTAL_PIP_CACHE_DIR_ARG__",
-                    Settings().PIP_CACHE_DIR_ARG,
-                ),
             ],
             script_dir=tmp_path,
             logger_name=__name__,
@@ -282,7 +252,7 @@ def test_templates_freeze(tmp_path, current_py_version):
         prefix="prefix",
     )
     dependencies_1 = _parse_pip_freeze_output(pip_freeze_venv_1)
-    assert "pip" in dependencies_1
+    assert "devtools" in dependencies_1
 
     # Write requirements file
     requirements_file = tmp_path / "requirements.txt"